Seperate application run part from original all in one script
This Dockerfile will create an image that will run the kuali student application
at a specific build.

It expects the database for the build to already be configured and available.

It will look for the ORACLE_DBA_URL to exist and point the DB at it.

If docker links in an oracle container (--link oracle:db) then it will
automatically construct a new jdbc url for that remote system.

Because of how the rice portal redirects the application needs to be aware of
the public ip it is running on.  It extracts its 172.17.0.0/16 IP address and
then modifies the startup configuration to use that value.

In order to connect to the application you will need to inspect the running
container and identify the IPAddress field value.

If the container is running in the boot2docker VM you may need to add a new
route inorder to be able to access the 172.17.0.0/16 network where the app is
running.